Face-Vision-3D_Pose/requirements.txt
torch>=0.4.1
torchvision>=0.2.1
numpy>=1.15.4
scipy>=1.1.0
matplotlib==3.0.2
dlib==19.9
opencv-python>=3.4.3.18
Flask
paho-mqtt
uwsgi